extract sizing and styling implicit functions
