+++ src/Parser.js
@@ -5,7 +5,6 @@ import functions from "./functions";
 import environments from "./environments";
 import MacroExpander from "./MacroExpander";
 import symbols from "./symbols";
-import utils from "./utils";
 import { validUnit } from "./units";
 import { supportedCodepoint } from "./unicodeScripts";
 import unicodeAccents from "./unicodeAccents";
@@ -435,17 +434,6 @@ export default class Parser {
         }
     }
 
-    // A list of the size-changing functions, for use in parseImplicitGroup
-    static sizeFuncs = [
-        "\\tiny", "\\sixptsize", "\\scriptsize", "\\footnotesize", "\\small",
-        "\\normalsize", "\\large", "\\Large", "\\LARGE", "\\huge", "\\Huge",
-    ];
-
-    // A list of the style-changing functions, for use in parseImplicitGroup
-    static styleFuncs = [
-        "\\displaystyle", "\\textstyle", "\\scriptstyle", "\\scriptscriptstyle",
-    ];
-
     // Old font functions
     static oldFontFuncs = {
         "\\rm": "mathrm",
@@ -527,25 +515,6 @@ export default class Parser {
                     endNameToken);
             }
             return result;
-        } else if (utils.contains(Parser.sizeFuncs, func)) {
-            // If we see a sizing function, parse out the implicit body
-            this.consumeSpaces();
-            const body = this.parseExpression(false, breakOnTokenText);
-            return new ParseNode("sizing", {
-                // Figure out what size to use based on the list of functions above
-                size: utils.indexOf(Parser.sizeFuncs, func) + 1,
-                value: body,
-            }, this.mode);
-        } else if (utils.contains(Parser.styleFuncs, func)) {
-            // If we see a styling function, parse out the implicit body
-            this.consumeSpaces();
-            const body = this.parseExpression(true, breakOnTokenText);
-            return new ParseNode("styling", {
-                // Figure out what style to use by pulling out the style from
-                // the function name
-                style: func.slice(1, func.length - 5),
-                value: body,
-            }, this.mode);
         } else if (func in Parser.oldFontFuncs) {
             const style = Parser.oldFontFuncs[func];
             // If we see an old font function, parse out the implicit body
@@ -576,7 +545,7 @@ export default class Parser {
             }, this.mode);
         } else {
             // Defer to parseGivenFunction if it's not a function we handle
-            return this.parseGivenFunction(start);
+            return this.parseGivenFunction(start, breakOnTokenText);
         }
     }
 
@@ -593,7 +562,10 @@ export default class Parser {
      * Same as parseFunction(), except that the base is provided, guaranteeing a
      * non-nullable result.
      */
-    parseGivenFunction(baseGroup: ParsedFuncOrArgOrDollar): ParseNode {
+    parseGivenFunction(
+        baseGroup: ParsedFuncOrArgOrDollar,
+        breakOnTokenText?: "]" | "}" | "$",
+    ): ParseNode {
         baseGroup = assertFuncOrArg(baseGroup);
         if (baseGroup.type === "fn") {
             const func = baseGroup.result;
@@ -611,7 +583,8 @@ export default class Parser {
 
             const {args, optArgs} = this.parseArguments(func, funcData);
             const token = baseGroup.token;
-            const result = this.callFunction(func, args, optArgs, token);
+            const result =
+                this.callFunction(func, args, optArgs, token, breakOnTokenText);
             return new ParseNode(result.type, result, this.mode);
         } else {
             return baseGroup.result;
@@ -626,11 +599,13 @@ export default class Parser {
         args: ParseNode[],
         optArgs: (?ParseNode)[],
         token?: Token,
+        breakOnTokenText?: "]" | "}" | "$",
     ): * {
         const context: FunctionContext = {
             funcName: name,
             parser: this,
             token,
+            breakOnTokenText,
         };
         const func = functions[name];
         if (func && func.handler) {
